# Implement prompt logprobs & Batched topk for computing logprobs