This PR:

- Added `prompt_logprobs` to `SamplingParams` and `RequestOutput`. This makes vLLM to support returning the log probabilities of prompt tokens, which is required to support `echo` in OpenAI server.
- Refactor the logprobs logic so that the query to `topk` logits is done in a batched fashion.

This PR will have merge conflicts with #1337. I think a good plan is to perform the optimization in #1337 along with the refactoring of `InputMetadata` after this PR is merged.

TODOs:

- [x] Test the correctness of this PR.
- [x] Test the performance with the new logprobs implementation.
- [ ] Maybe in a future PR: refactor bloated reference to `InputMetadata`