@wanmok 2023-10-16
> BTW, I got 2 logprobs per token when running `examples/llm_engine_example.py` where `prompt_logprobs=1`. Is this expected? It's a bit confusing because `logprobs=1` returns 1 log prob per token.
> 
> ```
> ..., prompt_logprobs=[None, {250: -3.594587802886963, 100: -1.414900302886963}, {9916: -8.579404830932617, 319: -3.582822799682617},  ...
> ```

A similar question here. Does the design support `logprobs=0`? In this case, we would like to only know the selected log probs in the prompt rather than top-k. This is required to implement the OpenAI API's `echo`.